server/requirements.txt
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.7
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.4
